chore(spec): little refacto on the spec to simplify further migration
